% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{dasl_logo}
\alias{dasl_logo}
\title{Add Fred Hutch DaSL logo}
\usage{
dasl_logo()
}
\value{
A list with ggplot2 components : \code{annotation_custom()}, \code{coord_cartesian()}, \code{theme()})
}
\description{
Logo placed at the bottom right corner of a ggplot
}
\examples{
\dontrun{
library(ggplot2)

ggplot(diamonds[sample(nrow(diamonds), 1000), ], aes(carat, price)) +
geom_point(aes(colour = color)) +
scale_colour_palette_d(hutch_palette) +
dasl_logo()
}
}
