% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutchpalette.R
\docType{data}
\name{hutch_palette}
\alias{hutch_palette}
\title{Fred Hutch Palette}
\format{
An object of class \code{palettes_palette} (inherits from \code{vctrs_list_of}, \code{vctrs_vctr}, \code{list}) of length 1.
}
\usage{
hutch_palette
}
\description{
The official colors of the Fred Hutchinson Cancer Center in Seattle, Washington.
}
\keyword{datasets}
