% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutchplot-package.R
\docType{package}
\name{hutchplot-package}
\alias{hutchplot}
\alias{hutchplot-package}
\title{hutchplot: Color Palette and Logos for the Fred Hutchinson Cancer Center}
\description{
Provides functions for ggplot2 color palettes and logos.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fhdsl/hutchplot}
  \item Report bugs at \url{https://github.com/fhdsl/hutchplot/issues}
}

}
\author{
\strong{Maintainer}: Howard Baek \email{howardbaek@fredhutch.org} (\href{https://orcid.org/0009-0000-8942-1618}{ORCID})

}
\keyword{internal}
