% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{dasl_logo}
\alias{dasl_logo}
\title{Add Fred Hutch DaSL logo}
\usage{
dasl_logo(x_unit = 25, y_unit = 20, height_unit = 12, width_unit = NULL)
}
\arguments{
\item{x_unit}{Numeric value indicating the amount of horizontal shifting.
Higher values shift towards the right, lower values shift towards the left}

\item{y_unit}{Numeric value indicating the amount of vertical shifting.
Greater values result in downward shifting, smaller values result in an
upward shifting.}

\item{height_unit}{Numeric value specifying width}

\item{width_unit}{Numeric value specifying width}
}
\value{
A list with ggplot2 components : \code{annotation_custom()},
\code{coord_cartesian()}, \code{theme()})
}
\description{
Logo placed at the bottom right corner of a ggplot
}
\examples{
\dontrun{
library(ggplot2)

ggplot(diamonds[sample(nrow(diamonds), 1000), ], aes(carat, price)) +
geom_point(aes(colour = color)) +
scale_colour_palette_d(hutch_palette) +
dasl_logo(x_unit = 25, y_unit = 10)
}
}
